pyproject.toml
[tool.poetry]
name = "pymandlebrot"
version = "0.0.0"
description = "A simple code to plot the Mandelbrot set"
authors = ["Freddy Wordingham <freddy@digilab.co.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.4"
pillow = "^10.3.0"
jupyter = "^1.0.0"
matplotlib = "^3.9.0"
jupyter-contrib-nbextensions = "^0.7.0"
tqdm = "^4.66.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
